requirements.txt
Babel==2.3.4
Django==1.9.12
Faker==0.7.7
Markdown==2.6.8
Pillow==4.0.0
Unidecode==0.04.20
django-appconf==1.0.2
django-compressor==2.1
django-extra-views==0.6.4
django-filter==1.0.1
django-haystack==2.5.1
django-mptt==0.8.7
django-oscar==1.3
django-tables2==1.0.7
django-treebeard==4.1.0
django-widget-tweaks==1.4.1
djangorestframework==3.5.4
factory-boy==2.8.1
mock==2.0.0
olefile==0.44
pbr==1.10.0
phonenumbers==7.7.5
purl==1.3
pycountry==17.1.8
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
sorl-thumbnail==12.4a1
